requirements.txt
requests>=2.25.1
google-auth-oauthlib>=0.4.6
google-auth>=1.35.0
google-api-python-client>=2.85.0
python-dotenv>=0.19.0